Rewrite editors into one component
