--- src/components/HelpersEditor/index.ts
@@ -1,1 +1,0 @@
-export * from './HelpersEditor';
