--- src/components/StylesEditor/index.ts
@@ -1,1 +1,0 @@
-export * from './StylesEditor';
