--- src/components/TextEditor/index.ts
@@ -1,1 +1,0 @@
-export * from './TextEditor';
