+++ src/components/index.ts
@@ -1,5 +1,3 @@
-export * from './HelpersEditor';
-export * from './StylesEditor';
 export * from './Text';
-export * from './TextEditor';
+export * from './CustomEditor';
 export * from './TextPanel';
